pyproject.toml
[tool.poetry]
name = "Spaceship_Titanic"
version = "0.1.0"
description = ""
authors = ["Yueren Xing <garrusxing@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
hydra-core = "~=1.3.2"
pydantic = "~=2.6.4"
dvc = {version = "~=3.49.0", extras=["gdrive", "gs"]}
mlflow = "~=2.11.3"
prefect = "~=2.16.9"
datachecks = "~=0.2.5"


numpy = "~=1.26.4"
pandas = "~=2.2.1"
tensorflow = "~=2.16.0"
tensorboard = "~=2.16.2"
scikit-learn = "~=1.4.1"
scikit-learn-extra = "~=0.3.0"
tqdm = "~=4.66.2"
mlxtend = "~=0.23.1"
joblib = "~=1.4.0"

fsspec = {version = "~=2024.3.1", extras=["gcs"]}
gcsfs = "~=2024.3.1"
google-cloud-secret-manager = "~=2.19.0"
google-cloud-compute = "~=1.18"
fastparquet = "~=2024.2.0"

dask = {version = "~=2024.4.1", extras=["complete"]}
dask-cloudprovider = {"version" = "~=2022.10.0", "extras" = ["gcp"]}
dask-ml = "~=2024.4.4"
distributed = "~=2024.4.1"
symspellpy = "~=6.7.7"

# ML models
xgboost = "2.0.3"
catboost = {extras = ["gpu"], version = "^1.2.5"}
lightgbm = "~=4.3.0"

# Automized EDA
ydata-profiling = "~=4.7.0"
autoviz = "~=0.1"

# Statistical Analysis
scipy = ">=1.4.1,<1.12"
statsmodels = "~=0.14.1"
MAPIE = "~=0.8.3"
distfit = "~=1.7.3"

# Imputation, Feature Engineering and Selection
Boruta = "~=0.3"
mrmr-selection = "~=0.2.8"
missingno = "~=0.5.2"
featuretools = "~=1.30.0"

# Optimization
parameter-sherpa = "~=1.0.6"
optuna = "~=3.6.1"
hgboost = "~=1.1.5"
dimod = "~=0.12.14"
dwave-system = "~=1.23.0"
dwave-scikit-learn-plugin  = "~=0.1.0"
deap = "~=1.4.1"

# Explainable-AI
shap = "~=0.45.0"
eli5 = "~=0.13.0"

# Visualization
matplotlib = "~=3.8.4"
seaborn = ">=0.10.1,<0.13"
plotly = "~=5.20.0"
bokeh = "~=3.4.1"

# Jupyter-related
jupyterlab = "~=4.1.5"
ipywidgets = "~=8.1.2"
jupyterlab-spellchecker = "~=0.8.4"
jupyterlab-code-formatter = "~=2.2.1"    
jupyterlab-git = "~=0.50.0"
JupyterLab-TensorBoard-Pro = "~=4.0.0"
# NLP
transformers = {version = "~=4.39.3", extras=["tensorflow"]}
tokenizers = "~=0.15.2"
psycopg2-binary = "~=2.9"
nltk = "~=3.8.1"
category-encoders = "^2.6.3"
pyod = "^1.1.3"
pygwalker = "^0.4.8.3"


[tool.poetry.group.dev.dependencies]
pytest = "~=8.1.1"
pytest-xdist = "~=3.5.0"
mock = "~=5.1.0"
black = "~=24.3.0"
isort = "~=5.13.2"
flake8 = "~=7.0.0"
mypy = "~=1.9.0"
loguru = "~=0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["SpaceshipTitanic"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.11"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
